requirements.txt
requests>=2.31.0
gunicorn>=21.2.0
flask>=3.0.0
pytz>=2023.3
supabase>=2.0.0
apscheduler>=3.10.0
pandas>=2.1.0
google-genai>=1.29.0
pypdf>=3.17.0
rich>=13.6.0
python-dotenv>=1.0.0
yfinance>=0.2.65
